+++ himan-plugins/source/precipitation.cpp
@@ -35,7 +35,7 @@ void precipitation::Process(std::shared_ptr<const plugin_configuration> conf)
 
 	// Get number of threads to use
 
-	unsigned short threadCount = ThreadCount(conf->ThreadCount());
+	short threadCount = ThreadCount(conf->ThreadCount());
 
 	if (conf->StatisticsEnabled())
 	{
@@ -148,7 +148,7 @@ void precipitation::Process(std::shared_ptr<const plugin_configuration> conf)
 	 * Each thread will have a copy of the target info.
 	 */
 
-	for (size_t i = 0; i < threadCount; i++)
+	for (short i = 0; i < threadCount; i++)
 	{
 
 		itsLogger->Info("Thread " + boost::lexical_cast<string> (i + 1) + " starting");
@@ -322,8 +322,8 @@ void precipitation::Calculate(shared_ptr<info> myTargetInfo,
 
 			shared_ptr<NFmiGrid> targetGrid(myTargetInfo->Grid()->ToNewbaseGrid());
 
-			int missingCount = 0;
-			int count = 0;
+			size_t missingCount = 0;
+			size_t count = 0;
 
 			bool equalGrids = (*myTargetInfo->Grid() == *RRInfo->Grid());
