+++ himan-plugins/source/tpot.cpp
@@ -49,7 +49,7 @@ void tpot::Process(std::shared_ptr<const plugin_configuration> conf)
 
 	// Get number of threads to use
 
-	unsigned short threadCount = ThreadCount(conf->ThreadCount());
+	short threadCount = ThreadCount(conf->ThreadCount());
 
 	if (conf->StatisticsEnabled())
 	{
@@ -173,7 +173,7 @@ void tpot::Process(std::shared_ptr<const plugin_configuration> conf)
 		aTimer->Start();
 	}
 	
-	for (size_t i = 0; i < threadCount; i++)
+	for (short i = 0; i < threadCount; i++)
 	{
 
 		itsLogger->Info("Thread " + boost::lexical_cast<string> (i + 1) + " starting");
